Implementação dos métodos GET e POST

+++ src/main/java/com/products/apirest/models/Product.java
@@ -2,6 +2,7 @@ package com.products.apirest.models;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -11,40 +12,46 @@ import javax.persistence.Table;
 
 @Entity
 @Table(name = "product")
-public class Product implements Serializable{
+public class Product implements Serializable {
 	private static final long serialVersionUID = 1L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private long id;
-	
+
 	private String name;
-	private BigDecimal quantity;
+	private BigInteger quantity;
 	private BigDecimal value;
-	
+
 	public long getId() {
 		return id;
 	}
+
 	public void setId(long id) {
 		this.id = id;
 	}
+
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
-	public BigDecimal getQuantity() {
+
+	public BigInteger getQuantity() {
 		return quantity;
 	}
-	public void setQuantity(BigDecimal quantity) {
+
+	public void setQuantity(BigInteger quantity) {
 		this.quantity = quantity;
 	}
+
 	public BigDecimal getValue() {
 		return value;
 	}
+
 	public void setValue(BigDecimal value) {
 		this.value = value;
 	}
-	
 }
